+++ Sources/XcodeServerCommon/CodeCoveragePreference.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 // TODO: Verify Order
-public enum CodeCoveragePreference: Int, Codable {
+public enum CodeCoveragePreference: Int16, Codable {
     case disabled = 0
     case enabled = 1
     case useSchemeSetting = 2
