+++ Sources/XcodeServerCommon/IssueStatus.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 /// The status of an integration issue.
-public enum IssueStatus: Int, Codable {
+public enum IssueStatus: Int16, Codable {
     case new = 1
     case unresolved = 2
     case resolved = 3
