+++ Sources/XcodeServerCommon/TriggerType.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 /// The type of trigger being executed.
-public enum TriggerType: Int, Codable {
+public enum TriggerType: Int16, Codable {
     case script = 1
     case email = 2
 }
